Update ripgrep setup
Separate installation and configuration. The former
fetches the latest release from github and installs
it locally. The latter provides the correct environment
(for any version of `rg` to run whatsoever).

Installing becomes a dedicated `setup.py` stage.
Configuration is a dependency of all shell setups.

This allows favorably installing `rg` from the
appropriate package manager, but preserves the
possibility of installing it locally (if ever necessary).

fixup! Update ripgrep setup